# Conflicts:
#	src/WebJobs.Script.WebHost/Middleware/FunctionInvocationMiddleware.cs
#	test/WebJobs.Script.Tests.Integration/WebHostEndToEnd/CSharpEndToEndTests.cs

+++ src/WebJobs.Script.WebHost/Scale/TableEntityConverter.cs
@@ -3,7 +3,7 @@
 
 using System;
 using System.Collections.Generic;
-using Microsoft.WindowsAzure.Storage.Table;
+using Microsoft.Azure.Cosmos.Table;
 using Newtonsoft.Json.Linq;
 
 namespace Microsoft.Azure.WebJobs.Script.WebHost.Scale
@@ -91,7 +91,9 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Scale
                     entityProperty = new EntityProperty(value.ToObject<Guid>());
                     return true;
                 case JTokenType.Integer:
-                    entityProperty = new EntityProperty(value.ToObject<int>());
+                    // to handle both ints and longs, we normalize integer values
+                    // to type long
+                    entityProperty = new EntityProperty(value.ToObject<long>());
                     return true;
                 case JTokenType.String:
                     entityProperty = new EntityProperty(value.ToObject<string>());
